Fixes #156 - Issue with spaces in paths
I've change up the way that EmberCLI::App#exec works so that it takes it's 'command' param as an array and then uses the splat operator to deconstruct it.
This offloads the work of escaping the various different paths for spaces (which is surprisingly a pain in the ass) onto the system/spawn calls.

Warning: This breaks logging out to the log path alongside stdout using the 'tee' command.
This is because the pipe command doesn't seem to be accepted by the system call when using the deconstructed arguments.
I'd like to fix this, but I don't have the time right now and that part isn't critical to me.