Change default `build_timeout` to 15 seconds
When handling views that use `include_ember_index_html`, the initial
request requires the `index.html` to be built and present.

The previous default value of `5` seconds isn't sufficient for the
request to kick off the build process and wait until completion.

To avoid confusion (and extra :arrows_counterclockwise: button clicks`),
we'll increase the timeout to `15` seconds.

Additionally, we'll expose the `EMBER_BUILD_TIMEOUT` environment
variable. Teams running on slower hardware can increase the timeout to a
better value without introducing git diff noise and churn.