Remove `build_timeout` configuration
The future use cases of this project expect that the build will complete
before serving the Ember applications.

Additionally, build timeouts used to act as a workaround for when
the EmberCLI build had errors that weren't being communicated to the
Rails process. Now that [EmberCLI writes errors to STDERR][stderr],
which EmberCLI-Rails [pipes STDERR to a file][pipe] for monitoring,
EmberCLI build errors will be raised by the Rails process.

For Users who care about timeouts, [rack-timeout] serves as a great
alternative.

[stderr]: https://github.com/ember-cli/ember-cli/pull/5039
[pipe]: https://github.com/thoughtbot/ember-cli-rails/pull/245
[rack-timeout]: https://github.com/heroku/rack-timeout