Introduce `mount_ember_assets` routing helper
Inspired by [#311].

The `mount_ember_app` helper mounts both a `Rack::File` application that
serves the EmberCLI-generated assets, as well as a catchall route (scoped
to the `to` argument) that redirects all requests to the Ember
application.

For users that have unconventional applications with Ember code embedded
in Rails pages, this is unwanted behavior.

The `mount_ember_assets` only mounts routes to serve the
EmberCLI-generated assets.

[#311]: https://github.com/thoughtbot/ember-cli-rails/issues/311#issuecomment-163739834