Don't mount routes to top-level context
Don't mount route helpers at top-level.

For instance, while serving an Ember application from `/admin/editor`,
the application could be mounted like this:

```rb
mount_ember_app :admin, to: "/admin/editor"
```

The previous implementation allows for this, as it was written with this
use case in mind.

Alternatively, the application could be mounted from within a
`namespace`, `scope`, or `constraint` block. Unfortunately, the previous
implementation doesn't support these use cases, since the routes are
mounted at the global level (i.e. in the first line of a
 `Rails.application.routes.draw` block).

This commit adds support for the latter case, taking into account the
surrounding context with which the route helpers are invoked.