Introduce `App#mountable?` and `App#to_rack`
For deploys that need to be mounted to be useful, [`File`][file] for
example, this commit exposes an API for encapsulating that logic.
`App#to_rack` will create an instance of a `Rack` application that can
be mounted via `mount $RACK_APP => "/path"`.

For deploys that don't serve assets from the file system,
[`Redis`][redis] for example, `mountable?` should return `false`.

[file]: https://github.com/thoughtbot/ember-cli-rails/blob/v0.7.0/lib/ember_cli/deploy/file.rb
[redis]: https://github.com/seanpdoyle/ember-cli-rails-deploy-redis