Fail `ember:compile` if build subprocess fails
Closes [#417][#417].

The following is the command `ember-cli-rails` generates for `rake
ember:compile`. It was generated using [the example application][repo].

```bash
/Users/seanpdoyle/src/ember-cli-rails-heroku-example/frontend/node_modules/.bin/ember build --environment 'development' --output-path '/Users/seanpdoyle/src/ember-cli-rails-heroku-example/tmp/ember-cli/apps/frontend' | /usr/bin/tee -a '/Users/seanpdoyle/src/ember-cli-rails-heroku-example/log/ember-frontend.development.log'
```

Pipes are streaming, so the pipe to `tee` will open before the first
program finishes writing, potentially ignoring a non-zero exit status.

The solution is to set the [`pipefail`][docs] option for the generated
sub-command.

[repo]: https://github.com/seanpdoyle/ember-cli-rails-heroku-example
[#417]: https://github.com/thoughtbot/ember-cli-rails/issues/417
[docs]: http://www.gnu.org/software/bash/manual/html_node/Pipelines.html