Remove `bin/heroku_install`
For deployments to Heroku, depend on `rake ember:{build,install}` being
executed as part of `rake assets:precompile`.

The removal of `bin/heroku_install` as part of a `postinstall` script in
the root-level `package.json` simplifies support as well.

This commit modifies the setup to incorporate checks for the presence of
`bower.json` to determine whether or not to cache the corresponding
`bower_components` directories, as well as the up-front installation of
the `bower` CLI tool.