add install_command_line_arguments to pass through to install commands
By default, yarn and npm will use updated versions of packages when possible.
This behavior may not be desirable in some cases, namely when using build systems
where you want to guarantee a specific version of your package is being used.

This change adds a new `install_command_line_arguments` option to the EmberCLI
initializer that can be used to customize the install command.