Drop support for End-of-Lifed Ruby versions
As of 20 June 2018, official support for [Ruby 2.2 has ended][2.2].

As of 31 March 2019, official support for [Ruby 2.3 has ended][2.3].

As of 5 April 2020, official support for [Ruby 2.4 has ended][2.4].

This commit configures the continuous integration test suite to execute
against `2.5` and `2.6`.

Additionally, this commit amends the `README` to mention `2.5` as the
minimally supported version.

[2.2]: https://www.ruby-lang.org/en/news/2018/06/20/support-of-ruby-2-2-has-ended/
[2.3]: https://www.ruby-lang.org/en/news/2019/03/31/support-of-ruby-2-3-has-ended/
[2.4]: https://www.ruby-lang.org/en/news/2020/04/05/support-of-ruby-2-4-has-ended/